[Chore] auth 관련 파일들 파일 구조 변경
