+++ src/features/auth/components/SignupCard.tsx
@@ -1,7 +1,7 @@
-import { mapRegisterError } from '@/apis/auth';
-import SignupAccountCard from '@/components/auth/SignupAccountCard';
-import SignupProfileCard from '@/components/auth/SignupProfileCard';
-import type { RegisterFormValues } from '@/types/auth';
+import SignupAccountCard from '@/features/auth/components/SignupAccountCard';
+import SignupProfileCard from '@/features/auth/components/SignupProfileCard';
+import type { RegisterFormValues } from '@/features/auth/types/auth';
+import { mapRegisterError } from '@/features/auth/utils/mapRegisterError';
 import React, { useState } from 'react';
 
 // 회원가입 단계 (1: 이메일 및 비밀번호, 2: 프로필 정보)
