+++ src/routes/Router.tsx
@@ -1,11 +1,11 @@
-import AgreementsPage from '@/pages/AgreementPage';
+import AgreementsPage from '@/features/auth/pages/AgreementPage';
+import LoginPage from '@/features/auth/pages/LoginPage';
+import SignupPage from '@/features/auth/pages/SignupPage';
 import ComplaintWizardPage from '@/pages/ComplaintWizardPage';
 import FaqPage from '@/pages/FaqPage';
-import LoginPage from '@/pages/LoginPage';
 import MainPage from '@/pages/MainPage';
 import MyComplaintsPage from '@/pages/MyComplaintsPage';
 import CrimeTypePage from '@/pages/PrecedentPage';
-import SignupPage from '@/pages/SignupPage';
 import { Route, Routes } from 'react-router-dom';
 
 // import SignupPage from '@/pages/SignupPage';
